Add passing poperties functionality to ProjectCard
I added functionality that allows for all aspects of each project card
to be fed to the card directly without anything hard-coded.
This includes:
- Project Title
- Project ID
- the Project Image
- a brief description
- an array of tags that include the name and color for each tag
- an array of links that include the name and href for each link

The only thing I haven't figured out yet is the Image. For some reason
when I bind the image to a variable, the image does not show. I will
look into this going forward.