Add navbar menu functionality for mobile
I added functionality that let the navbar work for mobile devices.
Previously the menu would not show from a mobile browser. Now, when the
navbar burger is clicked, the navbar menu is shown and the burger is
replaced by an X. When the X is clicked, the navbar menu disappears and
the burger is shown again.